backend/flask/requirements.txt
Flask==0.12.2
Flask-Migrate==2.1.0
Flask-SQLAlchemy==2.2
Flask-Script==2.0.5
Flask-RESTful==0.3.6
flask-cors==3.0.3
psycopg2==2.7.3
PyJWT==1.5.2
requests==2.18.4
uWSGI==2.0.15
